CLN: split to serial and parallel DAG
