+++ src/main/scala/io/github/facaiy/DAG/DAG.scala
@@ -1,0 +1,8 @@
+package io.github.facaiy.DAG
+
+import io.github.facaiy.DAG.core.{DAGNode, LazyCell}
+
+/**
+ * Created by facai on 6/2/17.
+ */
+trait DAG
