--- src/main/scala/io/github/facaiy/DAG/DAGNode.scala
@@ -1,82 +1,0 @@
-package io.github.facaiy.DAG
-
-import scala.concurrent.{Await, ExecutionContext, Future}
-import io.github.facaiy.DAG.LazyCell.{lazyCell, sequence}
-
-/**
- * Created by facai on 5/10/17.
- */
-sealed trait DAGNode[K, V] {
-  def name: K
-}
-
-object DAGNode {
-  def toLazyNetWork[K, V](nodes: Seq[DAGNode[K, V]]): Map[K, LazyCell[V]] = {
-    lazy val nodesMap: LazyCell[Map[K, LazyCell[V]]] =
-      lazyCell(nodes.map(toLazyCell).toMap)
-
-    def toLazyCell(n: DAGNode[K, V]): (K, LazyCell[V]) =
-      n match {
-        case InputNode(k, f) => k -> lazyCell(f())
-        case InternalNode(k, ds, f) =>
-          k -> nodesMap.flatMap { m =>
-            val inputs = ds.map(d => m.getOrElse(d, throw new NoSuchElementException(d.toString)))
-            sequence(inputs).map(f)
-          }
-      }
-
-    nodesMap.get()
-  }
-
-  def toFutureNetWork[K, V](nodes: Seq[DAGNode[K, V]])
-                           (implicit executor: ExecutionContext): Map[K, LazyCell[Future[V]]] = {
-
-    def toFutureCell(n: DAGNode[K, V]): DAGNode[K, Future[V]] =
-      n match {
-        case InputNode(k, f) =>
-          // TODO(facai), use `compose` to combine `f` and `Future.apply`
-          val g = () => Future(f())(executor)
-          InputNode(k, g)
-        case InternalNode(k, ds, f) =>
-          val g = (xs: Seq[Future[V]]) => Future.sequence(xs)(implicitly, executor).map(f)
-          InternalNode(k, ds, g)
-      }
-
-    toLazyNetWork(nodes.map(toFutureCell))
-  }
-
-  case class LazyFuture[A](lc: LazyCell[Future[A]]) {
-    import scala.concurrent.duration._
-
-    def getValue(timeOut: Int = 10): A =
-      Await.result(lc.get(), timeOut.seconds)
-  }
-
-  case class LazyCellOps[A](lc: LazyCell[A]) {
-    def getValue(): A = lc.get()
-  }
-}
-
-
-case class InputNode[K, V](override val name: K,
-                           transFunc: () => V) extends DAGNode[K, V]
-
-private case class InternalNode[K, V](override val name: K,
-                                      depends: Seq[K],
-                                      reduceFunc: Seq[V] => V) extends DAGNode[K, V]
-
-object ProcessNode {
-  def apply[K, V](name: K, depends: Seq[K], reduceFunc: Seq[V] => V): DAGNode[K, V] =
-    InternalNode(name, depends, reduceFunc)
-
-  def apply[K, V](name: K, depend: K, transFunc: V => V): DAGNode[K, V] =
-    apply(name, Seq(depend), (ss: Seq[V]) => transFunc(ss.head))
-}
-
-object OutputNode {
-  def apply[K, V](name: K, depends: Seq[K], reduceFunc: Seq[V] => Unit): DAGNode[K, V] =
-    InternalNode(name, depends, (ss: Seq[V]) => { reduceFunc(ss); ss.head })
-
-  def apply[K, V](name: K, depend: K, transFunc: V => Unit): DAGNode[K, V] =
-    apply(name, Seq(depend), (ss: Seq[V]) => transFunc(ss.head))
-}
