+++ src/main/scala/io/github/facaiy/DAG/core/LazyCell.scala
@@ -1,9 +1,4 @@
-package io.github.facaiy.DAG
-
-import scala.concurrent.Future
-import scala.language.implicitConversions
-
-import io.github.facaiy.DAG.DAGNode.{LazyCellOps, LazyFuture}
+package io.github.facaiy.DAG.core
 
 /**
  * Created by facai on 5/5/17.
@@ -29,8 +24,4 @@ object LazyCell {
   }
 
   def sequence[A](as: Seq[LazyCell[A]]): LazyCell[Seq[A]] = lazyCell(as.map(_.get()))
-
-  implicit def asLazyFuture[A](lc: LazyCell[Future[A]]): LazyFuture[A] = LazyFuture(lc)
-
-  implicit def asLazyCellOps[A](l: LazyCell[A]): LazyCellOps[A] = LazyCellOps(l)
 }
