+++ src/main/scala/io/github/facaiy/DAG/serial/DAG.scala
@@ -1,0 +1,34 @@
+package io.github.facaiy.DAG.serial
+
+import io.github.facaiy.DAG.{DAG => ParentDAG}
+import io.github.facaiy.DAG.core.{DAGNode, InputNode, InternalNode, LazyCell}
+import io.github.facaiy.DAG.core.LazyCell._
+
+/**
+ * Created by facai on 6/2/17.
+ */
+object DAG extends ParentDAG {
+  def toLazyNetWork[K, V](nodes: Seq[DAGNode[K, V]]): Map[K, LazyCell[V]] = {
+    lazy val nodesMap: LazyCell[Map[K, LazyCell[V]]] =
+      lazyCell(nodes.map(toLazyCell).toMap)
+
+    def toLazyCell(n: DAGNode[K, V]): (K, LazyCell[V]) =
+      n match {
+        case InputNode(k, f) => k -> lazyCell(f())
+        case InternalNode(k, ds, f) =>
+          k -> nodesMap.flatMap { m =>
+            val inputs = ds.map(d => m.getOrElse(d, throw new NoSuchElementException(d.toString)))
+            sequence(inputs).map(f)
+          }
+      }
+
+    nodesMap.get()
+  }
+
+  import scala.language.implicitConversions
+  implicit def asLazyCellOps[A](l: LazyCell[A]): LazyCellOps[A] = LazyCellOps(l)
+
+  case class LazyCellOps[A](lc: LazyCell[A]) {
+    def getValue(): A = lc.get()
+  }
+}
