+++ src/test/scala/io/github/facaiy/DAG/serial/DAGSuite.scala
@@ -1,13 +1,13 @@
-package io.github.facaiy.DAG
+package io.github.facaiy.DAG.serial
 
-import java.util.concurrent.{ConcurrentHashMap, Executors}
+import java.util.concurrent.ConcurrentHashMap
 
-import scala.concurrent.ExecutionContext
-
+import io.github.facaiy.DAG.core.{InputNode, LazyCell, OutputNode, ProcessNode}
+import io.github.facaiy.DAG.serial.DAG._
 import org.scalatest.FunSpec
 
 /**
- * Created by facai on 5/10/17.
+ * Created by facai on 6/2/17.
  */
 class DAGSuite extends FunSpec {
   describe("For lazy network,") {
@@ -55,7 +55,7 @@ class DAGSuite extends FunSpec {
 
       val nodes = Seq(input1, input2, input3, process1, process2, output1, output2)
 
-      val lm: Map[String, LazyCell[Int]] = DAGNode.toLazyNetWork(nodes)
+      val lm: Map[String, LazyCell[Int]] = DAG.toLazyNetWork(nodes)
       assert(lm.size === 7)
 
       lm("o1").getValue()
@@ -73,58 +73,6 @@ class DAGSuite extends FunSpec {
       assert(lm("p1").getValue() === 2)
       assert(lm("p2").getValue() === 5)
       assert(records.values().toArray().forall(_ === 1))
-    }
-  }
-
-  describe("For lazy and concurrent network") {
-    it("nodes are only be evaluated once.") {
-      import scala.concurrent.ExecutionContext.Implicits.global
-
-      val nodes = Seq(InputNode("input", () => System.currentTimeMillis()))
-      val fm = DAGNode.toFutureNetWork(nodes)
-
-      val before = fm("input").getValue()
-      Thread.sleep(1)
-      val after = fm("input").getValue()
-
-      assert(before === after)
-    }
-
-    it("nodes run in parallel.") {
-      implicit val ec = ExecutionContext.fromExecutor(Executors.newFixedThreadPool(2))
-
-      case class TimeRecord(id: String, start: Long, end: Long)
-
-      /* ------- helper method --------- */
-      def read(id: String, delayStart: Int = 10, runTime: Int = 1000)(): Seq[TimeRecord] = {
-        Thread.sleep(delayStart)
-        val start = System.currentTimeMillis()
-        Thread.sleep(runTime)
-        val end = System.currentTimeMillis()
-
-        List(TimeRecord(id, start, end))
-      }
-
-      def process(id: String)(st: Seq[Seq[TimeRecord]]): Seq[TimeRecord] =
-        read(id)() ++ st.reduce(_ ++ _)
-
-      /* ------- build network --------- */
-      val input1 = InputNode("input1", read("input1", 0))
-      val input2 = InputNode("input2", read("input2", 100))
-      val input3 = InputNode("input3", read("input3", 50))
-      val process1 = InternalNode(
-        "process1",
-        Seq("input1", "input2", "input3"),
-        process("process1"))
-      val nodes = Seq(input1, input2, input3, process1)
-
-      val fm = DAGNode.toFutureNetWork(nodes)
-
-      assert(
-        fm("process1").getValue()
-          .sortBy(_.start)
-          .sliding(2)
-          .exists(x => x(1).start < x(0).end))
     }
   }
 }
